# api/openapi.yaml
openapi: 3.1.0
info:
  title: Synveil API
  version: 0.1.0-blueprint
  summary: Authenticated metadata, Trash retention status, safe version restore, content-download, and exact-offset resumable upload API contract
  description: |
    This is the checked-in OpenAPI contract for the authenticated logical
    file/folder metadata, immutable version-history metadata, immutable
    content-download, resumable-upload, and metadata-only Trash retention slices.
    It records transport-wide conventions, foundation health routes, browser
    authentication transport, the first-run bootstrap boundary, owner-scoped
    library/root/node operations, immutable file-version listing and direct
    metadata lookup, security schemes, exact-offset raw-byte upload sessions,
    authenticated current and historical full/single-range content reads, and
    shared response shapes.

    Version-history listing is bounded, newest-first, node-scoped keyset
    pagination over immutable FileVersion rows. Direct version metadata uses
    the same opaque version ID accepted by the historical content route. Both
    reads expose only safe logical metadata and do not open an ObjectStore.

    Upload creation, owner-scoped status, bounded streaming append, retry-safe
    completion, and abort are implemented through the transport-neutral upload
    application service. Authenticated current-node and immutable-version
    content downloads are implemented through the transport-neutral content-read
...
servers:
...
paths:
...
  /api/v1/system/bootstrap-status:
    get:
      tags: [Bootstrap]
      operationId: getBootstrapStatus
      summary: Report whether first-run setup is still required
      x-synveil-status: IMPLEMENTED
      x-synveil-owner: Auth / Security, Rust Backend
      description: |
        Return only the minimum state needed to render first-run UX. The
        response contains no setup secret, password, database detail, or
        administrator identity and is always marked no-store. The route is
        intentionally bounded and has no distributed rate limiter in this
        phase; deployments must keep first-run exposure on a trusted/private
        or correctly terminated TLS network.
      security: []
      responses:
        '200':
          description: Current one-time bootstrap eligibility.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapStatusResponse'
        '503':
          $ref: '#/components/responses/BootstrapServiceUnavailable'
  /api/v1/bootstrap/admin:
    post:
      tags: [Bootstrap]
      operationId: createFirstAdmin
      summary: Create the first administrator and close setup
      x-synveil-status: IMPLEMENTED
      x-synveil-owner: Auth / Security, Rust Backend
      x-synveil-max-body-bytes: 16384
      description: |
        Create the first administrator through the existing race-safe
        bootstrap service. The request is strict, bounded JSON using the
        canonical login, login_key, and password fields. Browser requests are
        checked for same-origin provenance when Origin or Sec-Fetch-Site is
        present. No setup-secret field is part of the accepted contract, so
        deployment/network exposure is an explicit first-run trust assumption
        until a later installer or secret-gate contract is reviewed.

        A successful response reports setup_required=false and contains no
        password, verifier, session credential, or administrator record. This
        endpoint does not issue a browser session; the client must use the
        normal login route explicitly. Concurrent or repeated claims after the
        first successful transaction return bootstrap_closed and never reopen
        setup.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapAdminRequest'
      responses:
        '200':
          description: The first administrator was created and setup closed.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapStatusResponse'
        '400':
          $ref: '#/components/responses/BootstrapInvalidRequest'
        '413':
          $ref: '#/components/responses/BootstrapPayloadTooLarge'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/BootstrapClosed'
        '503':
          $ref: '#/components/responses/BootstrapServiceUnavailable'
...
  /api/v1/upload-sessions:
    post:
      tags: [Uploads]
      operationId: createUploadSession
      summary: Create an owner-scoped resumable upload session
      x-synveil-status: IMPLEMENTED
      x-synveil-owner: Storage / Uploads, Rust Backend
      x-synveil-max-json-body-bytes: 16384
      description: |
        Freeze either a CREATE_FILE destination or a REPLACE_CONTENT target and
        its canonical expected revision. The authenticated principal is the
        owner; no user_id is accepted from the client. Object keys, replica
        identifiers, staging handles, and physical paths are never returned.
      security:
        - BrowserSession: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadSessionRequest'
      responses:
        '201':
          description: The upload session was created with authoritative offset zero.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Upload-Offset:
              $ref: '#/components/headers/UploadOffset'
            Location:
              $ref: '#/components/headers/UploadLocation'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadSessionResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/UploadConflict'
        '413':
          $ref: '#/components/responses/UploadPayloadTooLarge'
        '507':
          $ref: '#/components/responses/CapacityUnavailable'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  parameters:
...
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      description: Session-bound CSRF proof issued by the authenticated CSRF endpoint; never use a query parameter.
      schema:
        type: string
        minLength: 128
        maxLength: 128
        pattern: '^[A-Fa-f0-9]{128}$'
  securitySchemes:
    BrowserSession:
      type: apiKey
      in: cookie
      name: synveil_session
      description: Opaque Secure/HttpOnly browser session; only its verifier is stored server-side.
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Scoped, rotatable, revocable device credential; never a password or URL parameter.
    PublicShareContext:
      type: apiKey
      in: cookie
      name: synveil_share
      description: Separate bounded public-share context; never promoted to a user session.
  headers:
    XRequestId:
      description: Trusted diagnostic request identifier echoed in the JSON envelope.
      schema:
        $ref: '#/components/schemas/OpaqueRequestId'
...
    RetryAfter:
      description: Seconds or HTTP date for bounded retry backoff.
      schema:
        type: string
...
    CacheControl:
      description: Sensitive authentication, bootstrap, and upload responses are not cacheable.
      schema:
        type: string
        const: no-store
...
    UploadOffset:
      description: Authoritative durable byte offset for this upload session.
      schema:
        $ref: '#/components/schemas/U64Decimal'
    UploadLocation:
      description: Canonical status/resume resource for the created upload session.
      schema:
        type: string
        pattern: '^/api/v1/upload-sessions/[0-9a-f-]{36}$'
  responses:
    InvalidRequest:
      description: The request body or bounded input is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UploadConflict:
      description: The upload state, expected length, target revision, or completion result conflicts with the requested operation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UploadPayloadTooLarge:
      description: The JSON command or aggregate raw-byte chunk exceeds its operation-specific limit.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    CapacityUnavailable:
      description: The upload cannot be admitted within current configured capacity.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: Authentication is required for the requested restricted operation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Request provenance or CSRF proof could not be verified.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: The process or a required operation dependency is unavailable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BootstrapInvalidRequest:
      description: The bounded first-run request body is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BootstrapClosed:
      description: First-run setup was already closed by an earlier successful claim.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BootstrapPayloadTooLarge:
      description: The first-run request body exceeds the 16 KiB transport limit.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BootstrapServiceUnavailable:
      description: The bootstrap state or first-administrator operation is unavailable.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Cache-Control:
          $ref: '#/components/headers/CacheControl'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CreateUploadSessionRequest:
      oneOf:
        - $ref: '#/components/schemas/CreateFileUploadRequest'
        - $ref: '#/components/schemas/ReplaceContentUploadRequest'
      discriminator:
        propertyName: operation
        mapping:
          CREATE_FILE: '#/components/schemas/CreateFileUploadRequest'
          REPLACE_CONTENT: '#/components/schemas/ReplaceContentUploadRequest'
    CreateFileUploadRequest:
      type: object
      additionalProperties: false
      required: [operation, library_id, parent_id, name, expected_bytes]
      properties:
        operation:
          type: string
          const: CREATE_FILE
        library_id:
          $ref: '#/components/schemas/OpaqueId'
        parent_id:
          $ref: '#/components/schemas/OpaqueId'
        name:
          type: string
          minLength: 1
          maxLength: 1024
          description: Raw logical name, never interpreted as an operating-system path.
        expected_bytes:
          $ref: '#/components/schemas/U64Decimal'
        expected_sha256:
          $ref: '#/components/schemas/Hash'
    ReplaceContentUploadRequest:
      type: object
      additionalProperties: false
      required: [operation, library_id, node_id, expected_revision, expected_bytes]
      properties:
        operation:
          type: string
          const: REPLACE_CONTENT
        library_id:
          $ref: '#/components/schemas/OpaqueId'
        node_id:
          $ref: '#/components/schemas/OpaqueId'
        expected_revision:
          $ref: '#/components/schemas/U64Decimal'
        expected_bytes:
          $ref: '#/components/schemas/U64Decimal'
        expected_sha256:
          $ref: '#/components/schemas/Hash'
    UploadSessionResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/UploadSessionResource'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    UploadSessionResource:
      type: object
      additionalProperties: false
      required: [id, type, attributes]
      properties:
        id:
          $ref: '#/components/schemas/OpaqueId'
        type:
          type: string
          const: upload_session
        attributes:
          $ref: '#/components/schemas/UploadSessionAttributes'
    UploadSessionAttributes:
      type: object
      additionalProperties: false
      required:
        - operation
        - state
        - received_bytes
        - expected_bytes
        - created_at
        - updated_at
        - expires_at
        - target
      properties:
        operation:
          type: string
          enum: [CREATE_FILE, REPLACE_CONTENT]
        state:
          $ref: '#/components/schemas/UploadSessionState'
        received_bytes:
          $ref: '#/components/schemas/U64Decimal'
        expected_bytes:
          $ref: '#/components/schemas/U64Decimal'
        expected_sha256:
          $ref: '#/components/schemas/Hash'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        target:
          $ref: '#/components/schemas/UploadTarget'
        last_error_code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
          description: Allow-listed stable diagnostic code only.
        terminal_failure_code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
          description: Allow-listed stable terminal failure code only.
        completion:
          $ref: '#/components/schemas/UploadCompletionResource'
    UploadSessionState:
      type: string
      enum: [OPEN, VERIFYING, COMMITTING, COMMITTED, FAILED, EXPIRED, ABORTED]
    UploadTarget:
      oneOf:
        - $ref: '#/components/schemas/CreateFileUploadTarget'
        - $ref: '#/components/schemas/ReplaceContentUploadTarget'
      discriminator:
        propertyName: operation
        mapping:
          CREATE_FILE: '#/components/schemas/CreateFileUploadTarget'
          REPLACE_CONTENT: '#/components/schemas/ReplaceContentUploadTarget'
    CreateFileUploadTarget:
      type: object
      additionalProperties: false
      required: [operation, library_id, parent_id, node_id, name]
      properties:
        operation:
          type: string
          const: CREATE_FILE
        library_id:
          $ref: '#/components/schemas/OpaqueId'
        parent_id:
          $ref: '#/components/schemas/OpaqueId'
        node_id:
          $ref: '#/components/schemas/OpaqueId'
        name:
          type: string
          minLength: 1
          maxLength: 1024
    ReplaceContentUploadTarget:
      type: object
      additionalProperties: false
      required: [operation, library_id, node_id, expected_revision]
      properties:
        operation:
          type: string
          const: REPLACE_CONTENT
        library_id:
          $ref: '#/components/schemas/OpaqueId'
        node_id:
          $ref: '#/components/schemas/OpaqueId'
        expected_revision:
          $ref: '#/components/schemas/U64Decimal'
...
    UploadCompletionResource:
      type: object
      additionalProperties: false
      required: [id, type, attributes]
      properties:
        id:
          $ref: '#/components/schemas/OpaqueId'
        type:
          type: string
          const: upload_completion
        attributes:
          $ref: '#/components/schemas/UploadCompletionAttributes'
    UploadCompletionAttributes:
      type: object
      additionalProperties: false
      required:
        - node_id
        - file_version_id
        - object_id
        - node_revision
        - bytes
        - sha256
        - committed_at
      properties:
        node_id:
          $ref: '#/components/schemas/OpaqueId'
        file_version_id:
          $ref: '#/components/schemas/OpaqueId'
        object_id:
          $ref: '#/components/schemas/OpaqueId'
        node_revision:
          $ref: '#/components/schemas/U64Decimal'
        bytes:
          $ref: '#/components/schemas/U64Decimal'
        sha256:
          $ref: '#/components/schemas/Hash'
        committed_at:
          type: string
          format: date-time
...
    BootstrapAdminRequest:
      type: object
      additionalProperties: false
      required: [login, login_key, password]
      properties:
        login:
          type: string
          minLength: 1
          maxLength: 1024
          description: Canonical user-visible login identifier.
        login_key:
          type: string
          minLength: 1
          maxLength: 1024
          description: Adapter-supplied canonical uniqueness key paired with login.
        password:
          type: string
          minLength: 1
          maxLength: 1024
          format: password
          description: Initial administrator password; never returned or logged.
...
    OpaqueRequestId:
      type: string
      minLength: 8
      maxLength: 128
      pattern: '^[A-Za-z0-9._~-]+$'
      description: Diagnostic identifier; not an idempotency key or credential.
    OpaqueId:
      type: string
      format: uuid
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      description: Opaque lowercase UUIDv7 public identifier.
    U64Decimal:
      type: string
      pattern: '^(0|[1-9][0-9]*)$'
      description: Non-negative integer serialized as canonical decimal text.
    Hash:
      type: string
      pattern: '^sha256:[0-9a-f]{64}$'
      description: Algorithm-qualified canonical integrity hash.
...
    BootstrapStatusResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: false
          required: [setup_required]
          properties:
            setup_required:
              type: boolean
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [request_id]
      properties:
        request_id:
          $ref: '#/components/schemas/OpaqueRequestId'
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
          description: Stable machine-readable error code from the reviewed registry.
        message:
          type: string
          minLength: 1
          maxLength: 512
          description: Safe human-readable message; clients must not branch on it.
        request_id:
          $ref: '#/components/schemas/OpaqueRequestId'
        retryable:
          type: boolean
        details:
          type: object
          description: Operation-specific allow-listed details; never stack traces, paths, secrets, or content.
          additionalProperties: true